Moving SearchField to its own folder
