+++ example/src/App.js
@@ -12,7 +12,12 @@ const App = () => {
 	return (
 		<div>
 			<h2>Autosuggest Field</h2>
-			<AutosuggestField endpoint="http://elasticpress.test/__elasticsearch/elasticpresstest-post-1/_doc/_search" />
+			<ElasticPressProvider
+				node="http://elasticpress.test/__elasticsearch"
+				indexName="elasticpresstest-post-1"
+			>
+				<AutosuggestField />
+			</ElasticPressProvider>
 
 			<h2>Related Posts</h2>
 			<RelatedContent wpApiRoot="http://elasticpress.test/wp-json" postId={2738} />
