--- src/components/AutosuggestField/AutosuggestField.test.js
@@ -1,7 +1,0 @@
-import AutosuggestField from './AutosuggestField';
-
-describe('AutosuggestField', () => {
-	it('is truthy', () => {
-		expect(AutosuggestField).toBeTruthy();
-	});
-});
