--- src/components/Posts/LoadMore.test.js
@@ -1,7 +1,0 @@
-import LoadMore from './LoadMore';
-
-describe('LoadMore', () => {
-	it('is truthy', () => {
-		expect(LoadMore).toBeTruthy();
-	});
-});
