--- src/components/Posts/PostItem.test.js
@@ -1,7 +1,0 @@
-import PostItem from './PostItem';
-
-describe('PostItem', () => {
-	it('is truthy', () => {
-		expect(PostItem).toBeTruthy();
-	});
-});
