--- src/components/Posts/Posts.test.js
@@ -1,7 +1,0 @@
-import Posts from './Posts';
-
-describe('Posts', () => {
-	it('is truthy', () => {
-		expect(Posts).toBeTruthy();
-	});
-});
