--- src/components/Posts/SearchField.test.js
@@ -1,7 +1,0 @@
-import SearchField from './SearchField';
-
-describe('SearchField', () => {
-	it('is truthy', () => {
-		expect(SearchField).toBeTruthy();
-	});
-});
