+++ src/components/SearchField/index.js
@@ -1,0 +1,2 @@
+export { default as SearchField } from './SearchField';
+export { default as query } from './query';
